app/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
passlib
PyJWT
pytest
httpx

